+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased] 
 - Show current script file name in window title
 - Add support for glob patterns when loading log files with `<`
-- add `require`, `exclude`, `requireToday`, and `clearFilter` commands to pre-filter lines as they are loaded.
+- Add `require`, `exclude`, `requireToday`, and `clearFilter` commands to pre-filter lines as they are loaded.
+- Save font size between runs
+- Correct error in column width calculation that led to log lines not filling the entire results pane
 
 ## [1.0.0] - 2019-10-15
 - First release